process continues

+++ src/main/java/com/gokhan/engcompany/service/EmployeeService.java
@@ -2,7 +2,6 @@ package com.gokhan.engcompany.service;
 
 import com.gokhan.engcompany.dto.EmployeeDto;
 import com.gokhan.engcompany.entity.Employee;
-import com.gokhan.engcompany.enums.DepartmentType;
 import com.gokhan.engcompany.repository.EmployeeRepository;
 import com.gokhan.engcompany.request.EmployeeRequest;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,7 +23,7 @@ public class EmployeeService {
 
         Employee manager = repository.findById(employeeId).get();
 
-        if ((manager.isManager() == true)) {
+        if ((manager.setManager() == true)) {
             return manager;
         } else { // to be added custom throw
             return null;
@@ -37,7 +36,7 @@ public class EmployeeService {
 
     public EmployeeDto getEmployeeDto(int employeeId) {
         Employee employee = getEmployeeEntity(employeeId);
-        if ( employee == null) {
+        if (employee == null) {
             return new EmployeeDto("This Employee is not exist");
         } else {
             return employee.toDto();
@@ -68,7 +67,7 @@ public class EmployeeService {
         Employee employee = new Employee();
         employee.setPerson(personService.insert(employeeRequest.personRequest));
         employee.setTitle(employeeRequest.title);
-        employee.isManager(employeeRequest.isManager);
+        employee.setManager(employeeRequest.isManager);
         employee = repository.save(employee);
         return employee.toDto();
     }
@@ -76,8 +75,8 @@ public class EmployeeService {
     public EmployeeDto promoteEmployee(EmployeeRequest employeeRequest, int employeeId) {
 
         Employee employee = getEmployeeEntity(employeeId);
-        if (!employee.isManager()) {
-            employee.isManager(employeeRequest.isManager);
+        if (!employee.setManager()) {
+            employee.setManager(employeeRequest.isManager);
             employee.setTitle(employeeRequest.title);
             return repository.save(employee).toDto();
         } else {
@@ -87,9 +86,13 @@ public class EmployeeService {
 
     public String deleteEmployee(int employeeId) {
 
-        if(getEmployeeEntity(employeeId)!= null) {
-            repository.deleteById(employeeId);
-            return "Kişi silindi";
+        if (getEmployeeEntity(employeeId) != null) {
+            if (getEmployeeEntity(employeeId).getDepartment() == null) {
+                repository.deleteById(employeeId);
+                return "Kişi silindi";
+            } else {
+                return "Kişiyi önce ilgili departmandan siliniz!";
+            }
         } else {
             return "Kişi silinemedi, çünkü yok";
         }
@@ -98,20 +101,33 @@ public class EmployeeService {
     public EmployeeDto updateEmployee(EmployeeRequest employeeRequest, int employeeId) {
         Employee employee = getEmployeeEntity(employeeId);
         employee.setPerson(personService.update(employeeRequest.personRequest, employee.getPerson().getPersonId()));
-        if(employee.isManager()!=employeeRequest.isManager && employee.isManager() == false) {
+        if (employee.setManager() != employeeRequest.isManager && employee.setManager() == false) {
             promoteEmployee(employeeRequest, employeeId);
         } else {
             employee.setTitle(employeeRequest.title);
-            employee.isManager(employeeRequest.isManager);
+            employee.setManager(employeeRequest.isManager);
         }
         return repository.save(employee).toDto();
     }
 
-    protected boolean checkIfEmployeeExists(int employeeId ) {
+    protected boolean checkIfEmployeeExists(int employeeId) {
         if (repository.existsByEmployeeId(employeeId)) {
             return true;
         } else {
             throw new EntityExistsException();
+        }
+    }
+
+    public String deleteManager(int managerId) {
+        if (getEmployeeEntity(managerId) != null) {
+            if (getEmployeeEntity(managerId).getDepartment() == null) {
+                repository.deleteById(managerId);
+                return "Manager silindi";
+            } else {
+                return "Manager önce ilgili departmandan siliniz!";
+            }
+        } else {
+            return "Manager silinemedi, çünkü yok";
         }
     }
 }
